examples/requirements.txt
-e .
pyyaml
aws_cdk.core
aws_cdk.aws_ec2
aws_cdk.aws_iam